// packages/www/src/types/playoffs.ts
export type NHLBracketPicksDto = {
  teamId: number;
  round: number;
  season: string;
  seriesCode: string;
}[] // z.infer<typeof NHLBracketPicksSchema>